ze03 2.7mm and 2.8mm
The ZE03 is an electrochemical gas detection module from Winsen, that
includes analog and digital sensor readouts.

Soldering the sensors is not advised, instead copper inserts/legs should
be soldered, and the sensor inserted in them.

Add support for kicad schematic, footprint and breakout for using a ZE03
sensor, that include 2.7mm and 2.8mm wholes for the legs (experimenting
with different tolerances, on paper both should work).